NEXT-31670 - Remove rules and rework product pricing

+++ src/Core/Checkout/Cart/CartRuleLoader.php
@@ -2,27 +2,24 @@
 
 namespace Shopware\Core\Checkout\Cart;
 
-use Doctrine\DBAL\Connection;
 use Psr\Log\LoggerInterface;
 use Shopware\Core\Checkout\Cart\Exception\CartTokenNotFoundException;
 use Shopware\Core\Checkout\Cart\LineItem\LineItem;
-use Shopware\Core\Checkout\Cart\Price\Struct\CartPrice;
-use Shopware\Core\Checkout\Cart\Tax\AbstractTaxDetector;
+use Shopware\Core\Checkout\Cart\Tax\CountryTaxCalculator;
 use Shopware\Core\Content\Rule\RuleCollection;
 use Shopware\Core\Content\Rule\RuleEntity;
 use Shopware\Core\Defaults;
 use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\DataAbstractionLayer\Exception\EntityNotFoundException;
+use Shopware\Core\Framework\Feature;
 use Shopware\Core\Framework\Log\Package;
-use Shopware\Core\Framework\Util\FloatComparator;
-use Shopware\Core\Framework\Uuid\Uuid;
 use Shopware\Core\Profiling\Profiler;
-use Shopware\Core\System\Country\CountryDefinition;
-use Shopware\Core\System\Country\CountryEntity;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Symfony\Contracts\Cache\CacheInterface;
 use Symfony\Contracts\Service\ResetInterface;
 
+/**
+ * @deprecated tag:v6.7.0 - #cache_rework_rule_reason#
+ */
 #[Package('checkout')]
 class CartRuleLoader implements ResetInterface
 {
@@ -31,29 +28,25 @@ class CartRuleLoader implements ResetInterface
     private ?RuleCollection $rules = null;
 
     /**
-     * @var array<string, float>
-     */
-    private array $currencyFactor = [];
-
-    /**
      * @internal
      */
     public function __construct(
-        private readonly AbstractCartPersister $cartPersister,
+        private readonly AbstractCartPersister $persister,
         private readonly Processor $processor,
         private readonly LoggerInterface $logger,
         private readonly CacheInterface $cache,
         private readonly AbstractRuleLoader $ruleLoader,
-        private readonly AbstractTaxDetector $taxDetector,
-        private readonly Connection $connection,
         private readonly CartFactory $cartFactory,
+        private readonly CountryTaxCalculator $taxCalculator
     ) {
     }
 
     public function loadByToken(SalesChannelContext $context, string $cartToken): RuleLoaderResult
     {
+        // todo@skroblin deprecate
+        // todo@skroblin upgrade guide
         try {
-            $cart = $this->cartPersister->load($cartToken, $context);
+            $cart = $this->persister->load($cartToken, $context);
 
             return $this->load($context, $cart, new CartBehavior($context->getPermissions()), false);
         } catch (CartTokenNotFoundException) {
@@ -70,17 +63,28 @@ class CartRuleLoader implements ResetInterface
 
     public function reset(): void
     {
+        if (Feature::isActive('cache_rework')) {
+            return;
+        }
         $this->rules = null;
     }
 
     public function invalidate(): void
     {
+        if (Feature::isActive('cache_rework')) {
+            return;
+        }
+
         $this->reset();
         $this->cache->delete(CachedRuleLoader::CACHE_KEY);
     }
 
     private function load(SalesChannelContext $context, Cart $cart, CartBehavior $behaviorContext, bool $new): RuleLoaderResult
     {
+        if (Feature::isActive('cache_rework')) {
+            return new RuleLoaderResult(cart: $cart, matchingRules: new RuleCollection());
+        }
+
         return Profiler::trace('cart-rule-loader', function () use ($context, $cart, $behaviorContext, $new) {
             $rules = $this->loadRules($context->getContext());
 
@@ -142,7 +146,7 @@ class CartRuleLoader implements ResetInterface
                 ++$iteration;
             } while ($recalculate);
 
-            $cart = $this->validateTaxFree($context, $cart, $behaviorContext);
+            $cart = $this->taxCalculator->calculate($cart, $context, $behaviorContext);
 
             $index = 0;
             foreach ($rules as $rule) {
@@ -157,7 +161,7 @@ class CartRuleLoader implements ResetInterface
 
             // save the cart if errors exist, so the errors get persisted
             if ($cart->getErrors()->count() > 0 || $this->updated($cart, $timestamps)) {
-                $this->cartPersister->save($cart, $context);
+                $this->persister->save($cart, $context);
             }
 
             return new RuleLoaderResult($cart, $rules);
@@ -170,7 +174,9 @@ class CartRuleLoader implements ResetInterface
             return $this->rules;
         }
 
-        return $this->rules = $this->ruleLoader->load($context)->filterForContext();
+        return $this->rules = $this->ruleLoader
+            ->load($context)
+            ->filterForContext();
     }
 
     private function cartChanged(Cart $previous, Cart $current): bool
@@ -183,32 +189,6 @@ class CartRuleLoader implements ResetInterface
             || $previousLineItems->getKeys() !== $currentLineItems->getKeys()
             || $previousLineItems->getTypes() !== $currentLineItems->getTypes()
         ;
-    }
-
-    private function detectTaxType(SalesChannelContext $context, float $cartNetAmount = 0): string
-    {
-        $currency = $context->getCurrency();
-        $currencyTaxFreeAmount = $currency->getTaxFreeFrom();
-        $isReachedCurrencyTaxFreeAmount = $currencyTaxFreeAmount > 0 && $cartNetAmount >= $currencyTaxFreeAmount;
-
-        if ($isReachedCurrencyTaxFreeAmount) {
-            return CartPrice::TAX_STATE_FREE;
-        }
-
-        $country = $context->getShippingLocation()->getCountry();
-
-        $isReachedCustomerTaxFreeAmount = $country->getCustomerTax()->getEnabled() && $this->isReachedCountryTaxFreeAmount($context, $country, $cartNetAmount);
-        $isReachedCompanyTaxFreeAmount = $this->taxDetector->isCompanyTaxFree($context, $country) && $this->isReachedCountryTaxFreeAmount($context, $country, $cartNetAmount, CountryDefinition::TYPE_COMPANY_TAX_FREE);
-
-        if ($isReachedCustomerTaxFreeAmount || $isReachedCompanyTaxFreeAmount) {
-            return CartPrice::TAX_STATE_FREE;
-        }
-
-        if ($this->taxDetector->useGross($context)) {
-            return CartPrice::TAX_STATE_GROSS;
-        }
-
-        return CartPrice::TAX_STATE_NET;
     }
 
     /**
@@ -231,75 +211,5 @@ class CartRuleLoader implements ResetInterface
         }
 
         return \count($timestamps) !== $cart->getLineItems()->count();
-    }
-
-    private function isReachedCountryTaxFreeAmount(
-        SalesChannelContext $context,
-        CountryEntity $country,
-        float $cartNetAmount = 0,
-        string $taxFreeType = CountryDefinition::TYPE_CUSTOMER_TAX_FREE
-    ): bool {
-        $countryTaxFreeLimit = $taxFreeType === CountryDefinition::TYPE_CUSTOMER_TAX_FREE ? $country->getCustomerTax() : $country->getCompanyTax();
-        if (!$countryTaxFreeLimit->getEnabled()) {
-            return false;
-        }
-
-        $countryTaxFreeLimitAmount = $countryTaxFreeLimit->getAmount() / $this->fetchCurrencyFactor($countryTaxFreeLimit->getCurrencyId(), $context);
-
-        $currency = $context->getCurrency();
-
-        $cartNetAmount /= $this->fetchCurrencyFactor($currency->getId(), $context);
-
-        // currency taxFreeAmount === 0.0 mean currency taxFreeFrom is disabled
-        return $currency->getTaxFreeFrom() === 0.0 && FloatComparator::greaterThanOrEquals($cartNetAmount, $countryTaxFreeLimitAmount);
-    }
-
-    private function fetchCurrencyFactor(string $currencyId, SalesChannelContext $context): float
-    {
-        if ($currencyId === Defaults::CURRENCY) {
-            return 1;
-        }
-
-        $currency = $context->getCurrency();
-        if ($currencyId === $currency->getId()) {
-            return $currency->getFactor();
-        }
-
-        if (\array_key_exists($currencyId, $this->currencyFactor)) {
-            return $this->currencyFactor[$currencyId];
-        }
-
-        $currencyFactor = $this->connection->fetchOne(
-            'SELECT `factor` FROM `currency` WHERE `id` = :currencyId',
-            ['currencyId' => Uuid::fromHexToBytes($currencyId)]
-        );
-
-        if (!$currencyFactor) {
-            throw new EntityNotFoundException('currency', $currencyId);
-        }
-
-        return $this->currencyFactor[$currencyId] = (float) $currencyFactor;
-    }
-
-    private function validateTaxFree(SalesChannelContext $context, Cart $cart, CartBehavior $behaviorContext): Cart
-    {
-        $totalCartNetAmount = $cart->getPrice()->getPositionPrice();
-        if ($context->getTaxState() === CartPrice::TAX_STATE_GROSS) {
-            $totalCartNetAmount -= $cart->getLineItems()->getPrices()->getCalculatedTaxes()->getAmount();
-        }
-        $taxState = $this->detectTaxType($context, $totalCartNetAmount);
-        $previous = $context->getTaxState();
-        if ($taxState === $previous) {
-            return $cart;
-        }
-
-        $context->setTaxState($taxState);
-        $cart->setData(null);
-        $cart = $this->processor->process($cart, $context, $behaviorContext);
-        if ($previous !== CartPrice::TAX_STATE_FREE) {
-            $context->setTaxState($previous);
-        }
-
-        return $cart;
     }
 }
